requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pymongo==4.11.3
python-dotenv==1.0.1
setuptools==75.8.0
Werkzeug==3.1.3
wheel==0.45.1
flask_cors
flask_login
PyPDF2
cohere
Flask_Session

